requirements.txt
# Backend Requirements
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0
websockets==12.0
loguru==0.7.2

# AI and Processing
google-generativeai==0.3.2
# PyMuPDF==1.23.5  # Commented out due to Visual Studio dependency

# Frontend Requirements
streamlit==1.28.1
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
streamlit-chat==0.1.1
streamlit-aggrid==0.3.4.post3

# Common
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
python-dateutil==2.8.2
